Mushrooms and sprites
Added a funtioning mushroom along with with combining the big and small sprites together to have a more accurate swap between mario sizes after getting a mushroom. On top of this a death zone was added to spikes in the level to either kill small mario or turn big mario into small mario.