Many new features and interactables
Added in a crouch mechanic which allows you to A) crouch and B) go down pipes. Which leads to the next addition, pipes (2/3 done) which take Mario to sub areas with coins. Coins, again, are another new thing added which you can see the amount collected on the new HUD. Blocks were added for coins and mushrooms and turn to empty blocks when hit. On top of that, A few other things were added that were not mentioned here but are mostly just quality of life changes.