Final update (V 1.0.0)
This is the final update for my Mario game before the due date for my course. I may update in the future but either way I hope anyone who finds and plays this has a more than enjoyable time even if the game is not 100% accurate.

Thanks again,
Ethan Schaller